[SOIN] Isole le chargement de configuration
... pour simplifier la logique de création d'une application, et
expliciter sa dépendance sur la conifguration.
