+++ src/index.ts
@@ -1,5 +1,7 @@
 import {fabriqueApplication, demarre} from "./serveur";
 import {port, webhookIds} from "./adaptateurEnvironnement";
+import {recupereConfiguration} from "./configuration";
 
-const application = fabriqueApplication(webhookIds);
+const configuration = recupereConfiguration(webhookIds);
+const application = fabriqueApplication(configuration);
 demarre(application, port);
