--- src/types/types.d.ts
@@ -1,20 +1,0 @@
-export type Configuration = {
-    services: Service[];
-}
-
-type Service = {
-    id: string;
-    type: TypeService;
-    configuration: ConfigurationService;
-}
-
-type ConfigurationService = ConfigurationRedirectionWebhook;
-
-type ConfigurationRedirectionWebhook = {
-    idWebhookMattermost: string;
-    formatage: string;
-    condition?: string;
-    entreesSansInjectionsDeMarkdown?: string[];
-}
-
-type TypeService = 'redirectionWebhook';
